# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HomeHub household management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meal-plans:
    get:
      operationId: getMealPlans
      tags: [meals]
      summary: List meal plans
      parameters:
        - name: weekStart
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of meal plan entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealPlanEntry"
    post:
      operationId: createMealPlanEntry
      tags: [meals]
      summary: Create or update a meal plan entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealPlanEntryInput"
      responses:
        "201":
          description: Created meal plan entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlanEntry"
  /meal-plans/{id}:
    patch:
      operationId: updateMealPlanEntry
      tags: [meals]
      summary: Update rating or notes on a meal plan entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMealPlanEntryInput"
      responses:
        "200":
          description: Updated meal plan entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlanEntry"
        "404":
          description: Not found
    delete:
      operationId: deleteMealPlanEntry
      tags: [meals]
      summary: Delete a meal plan entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /meal-ratings:
    get:
      operationId: getMealRatings
      tags: [meals]
      summary: Get per-member ratings for a meal plan entry
      parameters:
        - name: mealPlanId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of per-member ratings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealRatingList"
    post:
      operationId: upsertMealRating
      tags: [meals]
      summary: Set (upsert) a per-member rating for a meal plan entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertMealRatingInput"
      responses:
        "200":
          description: Upserted rating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealRating"
  /meal-ratings/{id}:
    delete:
      operationId: deleteMealRating
      tags: [meals]
      summary: Remove a per-member rating
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    MealPlanEntry:
      type: object
      properties:
        id:
          type: string
        weekStart:
          type: string
          format: date
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        meal:
          type: string
        notes:
          type: string
          nullable: true
        rating:
          type: string
          enum: [love, ok, skip]
          nullable: true
        propertyId:
          type: string
      required:
        - id
        - weekStart
        - dayOfWeek
        - mealType
        - meal
        - propertyId
    MealRating:
      type: object
      properties:
        id:
          type: string
        mealPlanId:
          type: string
        memberId:
          type: string
        rating:
          type: string
          enum: [love, ok, skip]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - mealPlanId
        - memberId
        - rating
        - createdAt
    MealRatingList:
      type: array
      items:
        $ref: "#/components/schemas/MealRating"
    UpsertMealRatingInput:
      type: object
      properties:
        mealPlanId:
          type: string
        memberId:
          type: string
        rating:
          type: string
          enum: [love, ok, skip]
      required:
        - mealPlanId
        - memberId
        - rating
    CreateMealPlanEntryInput:
      type: object
      properties:
        weekStart:
          type: string
          format: date
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        meal:
          type: string
        notes:
          type: string
          nullable: true
        rating:
          type: string
          enum: [love, ok, skip]
          nullable: true
        propertyId:
          type: string
      required:
        - weekStart
        - dayOfWeek
        - mealType
        - meal
        - propertyId
    UpdateMealPlanEntryInput:
      type: object
      properties:
        rating:
          type: string
          enum: [love, ok, skip]
          nullable: true
        notes:
          type: string
          nullable: true
...